ForwardProp & Begin of BackwardsProp
Forward propegation finished with a temporary way of inputting values, inital weights and biases.

A start of the backwards propegation is made.

+++ core/src/com/mygdx/ann/neurons/Synapse.java
@@ -4,17 +4,17 @@ package com.mygdx.ann.neurons;
 
 public class Synapse {
     
-    private int weight;
+    private double weight;
     
     public Synapse() {
         weight=1;
     }
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
 
-    public void setWeight(int updatedweight) {
+    public void setWeight(double updatedweight) {
         weight=updatedweight;
     }
 }
